move jspnrpc version to a separate module
